+++ components/ADempiere/Form/ExpressShipment/index.vue
@@ -18,10 +18,20 @@ along with this program. If not, see <https:www.gnu.org/licenses/>.
   <div class="main-express-receipt">
     <el-card class="box-card">
       <div slot="header" class="clearfix-express-receipt">
-        <el-form ref="form-express-receipt" inline label-position="top">
-          <el-row :gutter="20">
+        <el-form
+          ref="form-express-receipt"
+          label-position="top"
+          class="field-from"
+          inline
+        >
+          <el-row :gutter="10">
             <el-col :span="8">
-              <el-form-item :label="$t('VBankStatementMatch.field.businessPartner')" class="front-item-receipt">
+              <el-form-item class="front-item-receipt">
+                <template slot="label" style="width: 450px;">
+                  {{ $t('VBankStatementMatch.field.businessPartner') }}
+                  <br>
+                  <br>
+                </template>
                 <el-select
                   v-model="currentBusinessPartners"
                   placeholder="Please Select Business Partner"
@@ -39,7 +49,12 @@ along with this program. If not, see <https:www.gnu.org/licenses/>.
               </el-form-item>
             </el-col>
             <el-col :span="8">
-              <el-form-item :label="$t('form.expressShipment.field.salesOrder')" class="front-item-receipt">
+              <el-form-item class="front-item-receipt">
+                <template slot="label" style="width: 450px;">
+                  {{ $t('form.expressShipment.field.salesOrder') }}
+                  <br>
+                  <br>
+                </template>
                 <el-select
                   v-model="salesOrder"
                   placeholder="Please Select Purchase Order"
@@ -64,9 +79,11 @@ along with this program. If not, see <https:www.gnu.org/licenses/>.
               >
                 <template slot="label" style="width: 450px;">
                   {{ $t('form.expressReceipt.field.productcode') }}
-                  <el-checkbox v-model="isQuantityFromOrderLine" style="float: right;">
-                    Cantidad Completa de la Linea
-                  </el-checkbox>
+                  <p style="margin: 0px;">
+                    <el-checkbox v-model="isQuantityFromOrderLine">
+                      Cantidad Completa de la Linea
+                    </el-checkbox>
+                  </p>
                 </template>
                 <el-autocomplete
                   ref="searchValue"
@@ -108,6 +125,8 @@ along with this program. If not, see <https:www.gnu.org/licenses/>.
         :border="true"
         fit
         highlight-current-row
+        style="min-height: 400px;"
+        class="table-form"
         @cell-click="editQuantity"
       >
         <el-table-column
@@ -155,26 +174,33 @@ along with this program. If not, see <https:www.gnu.org/licenses/>.
             type="primary"
             icon="el-icon-check"
             class="button-base-icon"
-            style="float: right; margin: 10px;"
+            style="float: right; margin-top: 10px;font-size: 28px;"
             :disabled="isEmptyValue(salesOrder) || isComplete"
             @click="visible = true"
           />
           <el-button
             type="danger"
             icon="el-icon-close"
-            style="float: right;margin-top: 10px;"
+            style="float: right; margin-top: 10px;font-size: 28px;"
             class="button-base-icon"
             @click="closeForm"
           />
           <el-button
             type="info"
             plain
-            style="float: right; margin-top: 10px;"
+            style="float: right; margin-top: 10px;font-size: 28px;"
             class="button-base-icon"
             @click="clearForm"
           >
             <svg-icon icon-class="layers-clear" />
           </el-button>
+          <el-button
+            type="success"
+            class="button-base-icon"
+            icon="el-icon-refresh-right"
+            style="float: right; margin-top: 10px;font-size: 28px;"
+            @click="refreshLine"
+          />
         </el-col>
       </el-row>
     </el-card>
@@ -514,6 +540,14 @@ export default defineComponent({
       visible.value = false
     }
 
+    function refreshLine() {
+      const { id, uuid } = store.getters.getCurrentReceipt
+      store.dispatch('listLine', {
+        shipmentId: id,
+        shipmentUuid: uuid
+      })
+    }
+
     /**
    * Watch
    */
@@ -555,7 +589,8 @@ export default defineComponent({
       // Action Panel Footer
       processShipment,
       closeForm,
-      clearForm
+      clearForm,
+      refreshLine
     }
   }
 })
@@ -573,9 +608,37 @@ export default defineComponent({
   }
 }
 </style>
-<style>
-.el-form-item--medium .el-form-item__label {
-  line-height: 36px;
-  width: 450px;
+<style lang="scss">
+.field-from {
+  .el-form-item--medium .el-form-item__label {
+    line-height: 36px;
+    width: 450px;
+    font-size: 18px;
+  }
+}
+.select-from {
+  .el-select-dropdown__item {
+    padding: 0 20px;
+    position: relative;
+    white-space: nowrap;
+    overflow: hidden;
+    text-overflow: ellipsis;
+    color: #606266;
+    height: 34px;
+    line-height: 34px;
+    -webkit-box-sizing: border-box;
+    box-sizing: border-box;
+    cursor: pointer;
+    font-size: 18px;
+  }
+}
+.table-form {
+  .el-table__header-wrapper {
+    font-size: 18px;
+  }
+  .el-table .cell {
+    font-size: 18px;
+    line-height: 22px;
+  }
 }
 </style>
